+++ packages/app/src/plugin-worker/storage.ts
@@ -1,6 +1,6 @@
 import './register.js'
 import { createIndexedDB_KVStorageBackend, createInMemoryKVStorageBackend } from '@masknet/shared-base'
-import { broadcastMessage } from './message.js'
+import { broadcastMessage } from './message-port.js'
 
 export const inMemoryStorageBackend = createInMemoryKVStorageBackend((key, value) => {
     broadcastMessage('inMemoryStorage', [key, value])
